requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
markdown
langchain-core
langchain-community
langchain-google-genai
chromadb
jinja2
python-dotenv
langchain-text-splitters
